Updated NET version, changed implementation to not give away to much of the solution
