+++ src/Challenge1/Models/ErrorViewModel.cs
@@ -1,9 +1,8 @@
-namespace Challenge1.Models
-{
-	public class ErrorViewModel
-	{
-		public string? RequestId { get; set; }
+namespace Challenge1.Models;
 
-		public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-	}
+public class ErrorViewModel
+{
+	public string? RequestId { get; set; }
+
+	public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
 }
