+++ src/Challenge1/Program.cs
@@ -6,21 +6,23 @@ var builder = WebApplication.CreateBuilder(args);
 // Add services to the container.
 builder.Services.AddControllersWithViews();
 
-builder.Services.AddDbContext<TheAppContext>(opt => opt.UseInMemoryDatabase(DbData.DbName));
+builder.Services.AddDbContext<ApplicationDbContext>(opt => opt.UseInMemoryDatabase(DbData.DbName));
 
 var app = builder.Build();
+
+// Seed the database
+using (var scope = app.Services.CreateScope())
+{
+    var context = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
+    DbData.AddMockData(context);
+}
 
 // Configure the HTTP request pipeline.
 if (!app.Environment.IsDevelopment())
 {
     app.UseExceptionHandler("/Home/Error");
-    // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
     app.UseHsts();
 }
-
-var theScope = app.Services.CreateScope();
-var context = theScope.ServiceProvider.GetRequiredService<TheAppContext>();
-DbData.AddMockData(context);
 
 app.UseHttpsRedirection();
 app.UseStaticFiles();
